refactor(core): extract TextField prop types to types.ts
Move ICommonTextFieldProps helper + TTextFieldProps union into types.ts.
Use named ChangeEvent. ICommonTextFieldProps stays exported (body uses it
in type casts). No runtime change.
